refactor(cli): relocate --project id validator to _cli_helpers (dedupe)
_FLOW_ID_RE/_validate_project_id were byte-identical between cli_image.py
and cli_video.py after the video --project change. Move them into
_cli_helpers.py (the established shared home for cli_image/cli_video
helpers) and have both modules import from there instead, with the
existing negative-import drift test extended to pin it.
